LPS-123758 Move class
