+++ modules/apps/staging/staging-taglib/src/main/java/com/liferay/staging/taglib/servlet/taglib/BaseCssTag.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.staging.taglib.servlet.taglib.base;
+package com.liferay.staging.taglib.servlet.taglib;
 
 import com.liferay.frontend.taglib.util.TagAccessor;
 import com.liferay.frontend.taglib.util.TagResourceHandler;
