+++ modules/apps/staging/staging-taglib/src/main/java/com/liferay/staging/taglib/servlet/taglib/CheckboxTag.java
@@ -16,7 +16,6 @@ package com.liferay.staging.taglib.servlet.taglib;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.staging.taglib.internal.servlet.ServletContextUtil;
-import com.liferay.staging.taglib.servlet.taglib.base.BaseCssTag;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.jsp.PageContext;
